Fix Diacritization dataset name and citation
