+++ llmebench/datasets/WikiNewsDiacritization.py
@@ -1,0 +1,47 @@
+from llmebench.datasets.dataset_base import DatasetBase
+
+
+class WikiNewsDiacritizationDataset(DatasetBase):
+    def __init__(self, **kwargs):
+        super(WikiNewsDiacritizationDataset, self).__init__(**kwargs)
+
+    def metadata():
+        return {
+            "language": "ar",
+            "citation": """@inproceedings{darwish-etal-2017-arabic,
+                title = "{A}rabic Diacritization: Stats, Rules, and Hacks",
+                author = "Darwish, Kareem  and
+                  Mubarak, Hamdy  and
+                  Abdelali, Ahmed",
+                booktitle = "Proceedings of the Third {A}rabic Natural Language Processing Workshop",
+                month = apr,
+                year = "2017",
+                address = "Valencia, Spain",
+                publisher = "Association for Computational Linguistics",
+                url = "https://aclanthology.org/W17-1302",
+                doi = "10.18653/v1/W17-1302",
+                pages = "9--17",
+            }""",
+        }
+
+    def get_data_sample(self):
+        return {
+            "input": "Original sentence",
+            "label": "Sentence with diacritized words",
+        }
+
+    def load_data(self, data_path, no_labels=False):
+        data = []
+
+        with open(data_path, "r") as fp:
+            for line_idx, line in enumerate(fp):
+                text, diacritized_text = line.split("\t")
+                data.append(
+                    {
+                        "input": text.strip(),
+                        "label": diacritized_text.strip(),
+                        "line_number": line_idx,
+                    }
+                )
+
+        return data
